ai-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
groq
tavily-python
pydantic==2.4.2
python-dotenv==1.0.0
supabase>=2.0.0
